Modularize Setting.tsx
